xtensa: xclib: support for newlib based Xtensa toolchain
When using the Xtensa toolchain, the libc support can be based on
either a very old newlib version or a newer actively maintained
clean room implementation.

This change provides the needed support for when the Xtensa
toolchain is based on newlib. It is implemented as part of the code
guarded by CONFIG_XTENSA_LIBC.

Signed-off-by: William Tambe <williamt@cadence.com>