xtensa: improve dispatching of IRQ# greater than 31
This is an improvement to:
abeccfec289 xtensa: support for more than 32 interrupts

For XCHAL_NUM_INTERRUPTS > 32, this change prevents un-necessary
calls to `_xtensa_handle_one_int*()` as we check each special
register for pending interrupts.

For example, if XCHAL_NUM_INTERRUPTS > 96, without this change,
for interrupt 126 in level 1, following will happend:

- Check INTERRUPT and INTENABLE,
	call _xtensa_handle_one_int1(0, irqs) and exit while-loop.
- Check INTERRUPT1 and INTENABLE1,
	call _xtensa_handle_one_int1(1, irqs) and exit while-loop.
- Check INTERRUPT2 and INTENABLE2,
	call _xtensa_handle_one_int1(2, irqs) and exit while-loop.
- Check INTERRUPT3 and INTENABLE3,
	call _xtensa_handle_one_int1(3, irqs) and call the
	interrupt handler for interrupt 126.

The first three calls to _xtensa_handle_one_int1() are not required
because the variable `irqs` is null; but with this change, we avoid
making the un-necessary calls to `_xtensa_handle_one_int1()` for
the first three stps and exit the while-loop quicker.

This change also avoid un-necessarily calling
`_xtensa_handle_one_int*()` when within the last while-loop of the
function `xtensa_int*_c()`, the expression `irqs ^= m;` makes the
variable `irqs` null.

Signed-off-by: William Tambe <williamt@cadence.com>